+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ENHANCEMENTS:
 * compute: add `variables` attribute in `yandex_compute_instance_group` resource and data source
 * managed-kubernetes: add `node_ipv4_cidr_mask_size` attribute to `yandex_kubernetes_cluster` resource and data source
+* managed-kubernetes: add `deploy_policy.max_unavailable` and `deploy_policy.max_expansion` attributes to `yandex_kubernetes_node_group` resource and data source
 
 ## 0.35.0 (March 31, 2020)
 FEATURES: